+++ frontend/openchat-agent/src/services/community/candid/idl.js
@@ -1368,7 +1368,6 @@ export const idlFactory = ({ IDL }) => {
   });
   const CommunityPermissions = IDL.Record({
     'create_public_channel' : CommunityPermissionRole,
-    'block_users' : CommunityPermissionRole,
     'update_details' : CommunityPermissionRole,
     'remove_members' : CommunityPermissionRole,
     'invite_users' : CommunityPermissionRole,
@@ -1549,7 +1548,6 @@ export const idlFactory = ({ IDL }) => {
   });
   const OptionalCommunityPermissions = IDL.Record({
     'create_public_channel' : IDL.Opt(CommunityPermissionRole),
-    'block_users' : IDL.Opt(CommunityPermissionRole),
     'update_details' : IDL.Opt(CommunityPermissionRole),
     'remove_members' : IDL.Opt(CommunityPermissionRole),
     'invite_users' : IDL.Opt(CommunityPermissionRole),